requirements.txt
# ═══════════════════════════════════════════════════════════════
# Playwright Test Scaffold - Dependencies
# ═══════════════════════════════════════════════════════════════

# ─────────────────────────────────────────────────────────────────
# Core Testing
# ─────────────────────────────────────────────────────────────────
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3

# ─────────────────────────────────────────────────────────────────
# Test Execution
# ─────────────────────────────────────────────────────────────────
pytest-xdist==3.5.0
pytest-rerunfailures>=12.0   # retry on failure

# ─────────────────────────────────────────────────────────────────
# Reporting
# ─────────────────────────────────────────────────────────────────
allure-pytest==2.13.2
pytest-html==4.1.1

# ─────────────────────────────────────────────────────────────────
# CLI & Console
# ─────────────────────────────────────────────────────────────────
click>=8.1.0
rich>=13.0.0

# ─────────────────────────────────────────────────────────────────
# Configuration
# ─────────────────────────────────────────────────────────────────
pyyaml>=6.0
python-dotenv==1.0.0




pydantic==2.5.0
requests==2.31.0